Two different functions created and commented
First, recordOutput function was created for the Server so it could log each completed job and gather it's output in one location. Second, updateJobListBackup function was created incase the Server crashes, each time the Job List is updated on the Server end it will update the text file. When the Server comes back online it will read the text file and gather all the previous jobs before the crash occurred.